requirements.txt
tensorflow
numpy
matplotlib
opencv-python
Pillow
scikit-learn